requirements.txt
google-ads==24.1.0
google-auth==2.28.0
google-auth-oauthlib==1.2.0
openai==1.14.0
python-dotenv==1.0.1
